// setup/arcgis-experience-builder-1.6/ArcGISExperienceBuilder/client/dist/widgets/arcgis/query/src/config.ts
export enum CreateToolType {
  Point = 'Point',
  Polyline = 'Polyline',
  Polygon = 'Polygon',
  Rectangle = 'Rectangle',
  Circle = 'Circle'
}
// ...
export enum SpatialFilterType {
  CurrentMapExtent = 'CurrentMapExtent',
  InteractiveDrawMode = 'InteractiveDrawMode',
  None = ''
}

export enum SpatialRelation {
  Intersect = 'intersects',
  Contain = 'contains',
  Cross = 'crosses',
  EnvelopeIntersect = 'envelope-intersects',
  IndexIntersect = 'index-intersects',
  Overlap = 'overlaps',
  Touch = 'touches',
  Within = 'within',
}

export interface SpatialFilterObj {
  geometry: any
  graphic?: __esri.Graphic
  layer?: __esri.GraphicsLayer
  clearAfterApply?: boolean
  relation?: SpatialRelation
  buffer?: { distance: number, unit: UnitType }
}
// ...
export enum UnitType {
  Miles = 'Miles',
  Kilometers = 'Kilometers',
  Feet = 'Feet',
  Meters = 'Meters',
  Yards = 'Yards',
  NauticalMiles = 'NauticalMiles',
}